+++ lib/myrrha/ext/domain.rb
@@ -2,16 +2,19 @@ module Domain
   module CoercionMethods
 
     def coercions(&bl)
-      @coercions ||= ::Myrrha::Coercions.new{|c| c.main_target_domain = self}
+      @coercions ||= ::Myrrha::Coercions.new{|c|
+        c.main_target_domain = self
+        c.error_handler = lambda{|v,t,c|
+          raise c if TypeError===c
+          domain_error!(v)
+        }
+      }
       @coercions.append(&bl) if bl
       @coercions
     end
 
     def coerce(arg)
       coercions.coerce(arg, self)
-    rescue Myrrha::Error => ex
-      raise ex.cause if ex.cause
-      domain_error!(arg)
     end
     alias_method :[], :coerce